olx_selenium_parser/requirements.txt
aiohttp
beautifulsoup4
python-dotenv
selenium